This following model is used for classification of  types of crops to grow in a field based on its biochemical and biophysical factors .
The model  is designed on Random Forest Classifier and is able to obtain ~98% accuracy. The data can classify 22 types of crop based on the factors of soil and weather to provide high yield in agricultural .

1.ideation: There are many  places where water is scarce, Somewhere there are not much nutrient value in the soil.I decided to help the Farmers using this model as they can grow crops based on the natural condition and can achieve high yield.

2.Scope: The Project can be used in agricultural field to increase type of crop growth and promote agriculture more.

3.Data Collection: For this  project I collected my data from Kaggle. The following data is uploaded by ATHARVA INGLE. The following dataset contains 7 features on which the classification is based.
Data_link:https://www.kaggle.com/datasets/atharvaingle/crop-recommendation-dataset

4.Model overview and Performance:
The model is based on RandomForest
Classifier and able achieve 98% accuracy using the model. Upon Testing I've found that it classifies the out put correctly most of the times.